refactor(metrics): move reporter SPI to runtime/service; fail-on-invoke fallback (PR0 review)
Address review feedback from @dimas-b on #5068:

- Move IcebergMetricsReporter, MetricsReportEnvelope and MetricType from
  polaris-core to runtime/service (org.apache.polaris.service.metrics). With
  metrics detached from BasePersistence, core no longer needs the reporter SPI;
  only runtime/service and the reporter extensions use it. The base and JDBC
  reporter extensions now depend on polaris-runtime-service (same pattern as
  extensions/semantic-models).
- ServiceProducers.metricsReporter no longer logs a WARN at startup when no
  reporter matches the configured type (a false alarm for downstream builds that
  intentionally omit a reporter). It now returns a reporter that logs an error
  and throws only if reportMetric() is actually invoked.
