+++ extensions/metrics-reports/base/src/main/java/org/apache/polaris/extension/metrics/reports/LoggingMetricsReporter.java
@@ -22,8 +22,8 @@ import com.google.common.annotations.VisibleForTesting;
 import io.smallrye.common.annotation.Identifier;
 import jakarta.enterprise.context.ApplicationScoped;
 import java.util.function.Consumer;
-import org.apache.polaris.core.metrics.IcebergMetricsReporter;
-import org.apache.polaris.core.metrics.MetricsReportEnvelope;
+import org.apache.polaris.service.metrics.IcebergMetricsReporter;
+import org.apache.polaris.service.metrics.MetricsReportEnvelope;
 import org.jspecify.annotations.NonNull;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
