+++ extensions/metrics-reports/persistence/relational-jdbc/src/main/java/org/apache/polaris/extension/metrics/jdbc/PersistingMetricsReporter.java
@@ -28,11 +28,11 @@ import org.apache.iceberg.metrics.CommitReport;
 import org.apache.iceberg.metrics.ScanReport;
 import org.apache.polaris.core.auth.PolarisPrincipal;
 import org.apache.polaris.core.context.RequestIdSupplier;
-import org.apache.polaris.core.metrics.IcebergMetricsReporter;
-import org.apache.polaris.core.metrics.MetricsReportEnvelope;
 import org.apache.polaris.core.persistence.metrics.CommitMetricsRecord;
 import org.apache.polaris.core.persistence.metrics.MetricsPersistence;
 import org.apache.polaris.core.persistence.metrics.ScanMetricsRecord;
+import org.apache.polaris.service.metrics.IcebergMetricsReporter;
+import org.apache.polaris.service.metrics.MetricsReportEnvelope;
 import org.jspecify.annotations.NonNull;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
